requirements.txt
pandas
pandassql
SQLAlchemy=1.4.46
mysqlclient
python-dotenv